refactor: Split out old/new transit filter mapping in separate classes
