+++ application/src/main/java/org/opentripplanner/apis/transmodel/mapping/TransitFilterOldWayMapper.java
@@ -12,14 +12,15 @@ import org.opentripplanner.apis.transmodel.support.DataFetcherDecorator;
 import org.opentripplanner.apis.transmodel.support.GqlUtil;
 import org.opentripplanner.routing.api.request.request.TransitRequestBuilder;
 import org.opentripplanner.routing.api.request.request.filter.SelectRequest;
-import org.opentripplanner.routing.api.request.request.filter.TransitFilter;
-import org.opentripplanner.routing.api.request.request.filter.TransitFilterRequest;
 import org.opentripplanner.transit.model.basic.MainAndSubMode;
 import org.opentripplanner.transit.model.basic.SubMode;
 import org.opentripplanner.transit.model.basic.TransitMode;
 import org.opentripplanner.transit.model.framework.FeedScopedId;
 
-class FilterMapper {
+class TransitFilterOldWayMapper {
+
+  /** This is a utility class, only static methods */
+  private TransitFilterOldWayMapper() {}
 
   @SuppressWarnings("unchecked")
   static void mapFilterOldWay(
@@ -145,30 +146,5 @@ class FilterMapper {
       .peek(it -> it.withTransportModes(tModes))
       .map(it -> it.build())
       .toList();
-  }
-
-  @SuppressWarnings("unchecked")
-  static List<TransitFilter> mapFilterNewWay(List<Map<String, ?>> filters) {
-    var filterRequests = new ArrayList<TransitFilter>();
-
-    for (var filterInput : filters) {
-      var filterRequestBuilder = TransitFilterRequest.of();
-
-      if (filterInput.containsKey("select")) {
-        for (var selectInput : (List<Map<String, List<?>>>) filterInput.get("select")) {
-          filterRequestBuilder.addSelect(SelectRequestMapper.mapSelectRequest(selectInput));
-        }
-      }
-
-      if (filterInput.containsKey("not")) {
-        for (var selectInput : (List<Map<String, List<?>>>) filterInput.get("not")) {
-          filterRequestBuilder.addNot(SelectRequestMapper.mapSelectRequest(selectInput));
-        }
-      }
-
-      filterRequests.add(filterRequestBuilder.build());
-    }
-
-    return filterRequests;
   }
 }
